Rename class and extract inner method for click handler.
