+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.PermissionsApp"
         tools:ignore="AllowBackup">
-        <activity android:name=".Display_apps" />
+        <activity android:name=".DisplayApps" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
